Move KeyRing::Impl to its own file
