+++ zypp/KeyRing.cc
@@ -9,6 +9,9 @@
 /** \file	zypp/KeyRing.cc
  *
 */
+
+#include "zypp_detail/keyring_p.h"
+
 #include <iostream>
 #include <fstream>
 #include <optional>
@@ -27,7 +30,6 @@
 #include <zypp/base/Gettext.h>
 #include <zypp-core/fs/WatchFile>
 #include <zypp/PathInfo.h>
-#include <zypp/KeyRing.h>
 #include <zypp/ExternalProgram.h>
 #include <zypp/TmpPath.h>
 #include <zypp/ZYppCallbacks.h>       // JobReport::instance
@@ -110,231 +112,74 @@ namespace zypp
     report( data );
   }
 
-  namespace
-  {
-    ///////////////////////////////////////////////////////////////////
-    /// \class CachedPublicKeyData
-    /// \brief Functor returning the keyrings data (cached).
-    /// \code
-    ///   const std::list<PublicKeyData> & cachedPublicKeyData( const Pathname & keyring );
-    /// \endcode
-    ///////////////////////////////////////////////////////////////////
-    struct CachedPublicKeyData : private base::NonCopyable
-    {
-      const std::list<PublicKeyData> & operator()( const Pathname & keyring_r ) const
-      { return getData( keyring_r ); }
+  ///////////////////////////////////////////////////////////////////
 
-      void setDirty( const Pathname & keyring_r )
-      { _cacheMap[keyring_r].setDirty(); }
+  CachedPublicKeyData::Manip::Manip(CachedPublicKeyData &cache_r, filesystem::Pathname keyring_r)
+    : _cache { cache_r }
+    , _keyring { std::move(keyring_r) }
+  {}
 
-      ///////////////////////////////////////////////////////////////////
-      /// \brief Helper providing on demand a KeyManagerCtx to manip the cached keyring.
-      ///
-      /// The 1st call to \ref keyManagerCtx creates the \ref KeyManagerCtx. Returning
-      /// the context tags the cached data as dirty. Should be used to import/delete keys
-      /// in a cache keyring.
-      struct Manip {
-        NON_COPYABLE_BUT_MOVE( Manip );
-        Manip( CachedPublicKeyData & cache_r, Pathname keyring_r )
-        : _cache { cache_r }
-        , _keyring { std::move(keyring_r) }
-        {}
-
-        KeyManagerCtx & keyManagerCtx() {
-          if ( not _context ) {
-            _context = KeyManagerCtx::createForOpenPGP( _keyring );
-          }
-          // frankly: don't remember why an explicit setDirty was introduced and
-          // why WatchFile was not enough. Maybe some corner case when the keyrings
-          // are created?
-          _cache.setDirty( _keyring );
-          return _context.value();
-        }
-
-      private:
-        CachedPublicKeyData & _cache;
-        Pathname _keyring;
-        std::optional<KeyManagerCtx> _context;
-      };
-      ///////////////////////////////////////////////////////////////////
-
-      /** Helper providing on demand a KeyManagerCtx to manip the cached keyring. */
-      Manip manip( Pathname keyring_r ) { return Manip( *this, std::move(keyring_r) ); }
-
-    private:
-      struct Cache
-      {
-        Cache() {}
-
-        void setDirty()
-        {
-          _keyringK.reset();
-          _keyringP.reset();
-        }
-
-        void assertCache( const Pathname & keyring_r )
-        {
-          // .kbx since gpg2-2.1
-          if ( !_keyringK )
-            _keyringK.reset( new WatchFile( keyring_r/"pubring.kbx", WatchFile::NO_INIT ) );
-          if ( !_keyringP )
-            _keyringP.reset( new WatchFile( keyring_r/"pubring.gpg", WatchFile::NO_INIT ) );
-        }
-
-        bool hasChanged() const
-        {
-          bool k = _keyringK->hasChanged();	// be sure both files are checked
-          bool p = _keyringP->hasChanged();
-          return k || p;
-        }
-
-        std::list<PublicKeyData> _data;
-
-      private:
-
-        scoped_ptr<WatchFile> _keyringK;
-        scoped_ptr<WatchFile> _keyringP;
-      };
-
-      typedef std::map<Pathname,Cache> CacheMap;
-
-      const std::list<PublicKeyData> & getData( const Pathname & keyring_r ) const
-      {
-        Cache & cache( _cacheMap[keyring_r] );
-        // init new cache entry
-        cache.assertCache( keyring_r );
-        return getData( keyring_r, cache );
-      }
-
-      const std::list<PublicKeyData> & getData( const Pathname & keyring_r, Cache & cache_r ) const
-      {
-        if ( cache_r.hasChanged() ) {
-          cache_r._data = KeyManagerCtx::createForOpenPGP( keyring_r ).listKeys();
-          MIL << "Found keys: " << cache_r._data  << endl;
-        }
-        return cache_r._data;
-      }
-
-      mutable CacheMap _cacheMap;
-    };
-    ///////////////////////////////////////////////////////////////////
-
-
+  KeyManagerCtx &CachedPublicKeyData::Manip::keyManagerCtx() {
+    if ( not _context ) {
+      _context = KeyManagerCtx::createForOpenPGP( _keyring );
+    }
+    // frankly: don't remember why an explicit setDirty was introduced and
+    // why WatchFile was not enough. Maybe some corner case when the keyrings
+    // are created?
+    _cache.setDirty( _keyring );
+    return _context.value();
   }
 
-  ///////////////////////////////////////////////////////////////////
-  //
-  //	CLASS NAME : KeyRing::Impl
-  //
-  /** KeyRing implementation. */
-  struct KeyRing::Impl
+  CachedPublicKeyData::Cache::Cache() {}
+
+  void CachedPublicKeyData::Cache::setDirty()
   {
-    Impl( const Pathname & baseTmpDir )
-    : _trusted_tmp_dir( baseTmpDir, "zypp-trusted-kr" )
-    , _general_tmp_dir( baseTmpDir, "zypp-general-kr" )
-    , _base_dir( baseTmpDir )
-    {
-      MIL << "Current KeyRing::DefaultAccept: " << _keyRingDefaultAccept << endl;
+    _keyringK.reset();
+    _keyringP.reset();
+  }
+
+  void CachedPublicKeyData::Cache::assertCache(const filesystem::Pathname &keyring_r)
+  {
+    // .kbx since gpg2-2.1
+    if ( !_keyringK )
+      _keyringK.reset( new WatchFile( keyring_r/"pubring.kbx", WatchFile::NO_INIT ) );
+    if ( !_keyringP )
+      _keyringP.reset( new WatchFile( keyring_r/"pubring.gpg", WatchFile::NO_INIT ) );
+  }
+
+  bool CachedPublicKeyData::Cache::hasChanged() const
+  {
+    bool k = _keyringK->hasChanged();	// be sure both files are checked
+    bool p = _keyringP->hasChanged();
+    return k || p;
+  }
+
+  const std::list<PublicKeyData> &CachedPublicKeyData::operator()(const filesystem::Pathname &keyring_r) const
+  { return getData( keyring_r ); }
+
+  void CachedPublicKeyData::setDirty(const filesystem::Pathname &keyring_r)
+  { _cacheMap[keyring_r].setDirty(); }
+
+  CachedPublicKeyData::Manip CachedPublicKeyData::manip(filesystem::Pathname keyring_r) { return Manip( *this, std::move(keyring_r) ); }
+
+  const std::list<PublicKeyData> &CachedPublicKeyData::getData(const filesystem::Pathname &keyring_r) const
+  {
+    Cache & cache( _cacheMap[keyring_r] );
+    // init new cache entry
+    cache.assertCache( keyring_r );
+    return getData( keyring_r, cache );
+  }
+
+  const std::list<PublicKeyData> &CachedPublicKeyData::getData(const filesystem::Pathname &keyring_r, Cache &cache_r) const
+  {
+    if ( cache_r.hasChanged() ) {
+      cache_r._data = KeyManagerCtx::createForOpenPGP( keyring_r ).listKeys();
+      MIL << "Found keys: " << cache_r._data  << std::endl;
     }
+    return cache_r._data;
+  }
 
-    void importKey( const PublicKey & key, bool trusted = false );
-    void multiKeyImport( const Pathname & keyfile_r, bool trusted_r = false );
-    void deleteKey( const std::string & id, bool trusted );
 
-    std::string readSignatureKeyId( const Pathname & signature );
-
-    bool isKeyTrusted( const std::string & id )
-    { return bool(publicKeyExists( id, trustedKeyRing() )); }
-    bool isKeyKnown( const std::string & id )
-    { return publicKeyExists( id, trustedKeyRing() ) || publicKeyExists( id, generalKeyRing() ); }
-
-    std::list<PublicKey> trustedPublicKeys()
-    { return publicKeys( trustedKeyRing() ); }
-    std::list<PublicKey> publicKeys()
-    { return publicKeys( generalKeyRing() ); }
-
-    const std::list<PublicKeyData> & trustedPublicKeyData()
-    { return publicKeyData( trustedKeyRing() ); }
-    const std::list<PublicKeyData> & publicKeyData()
-    { return publicKeyData( generalKeyRing() ); }
-
-    void dumpPublicKey( const std::string & id, bool trusted, std::ostream & stream )
-    { dumpPublicKey( id, ( trusted ? trustedKeyRing() : generalKeyRing() ), stream ); }
-
-    PublicKey exportPublicKey( const PublicKeyData & keyData )
-    { return exportKey( keyData, generalKeyRing() ); }
-    PublicKey exportTrustedPublicKey( const PublicKeyData & keyData )
-    { return exportKey( keyData, trustedKeyRing() ); }
-
-    bool verifyFileSignatureWorkflow( keyring::VerifyFileContext & context_r )
-    {
-      // Assert result and return value are in sync
-      context_r.fileAccepted( _verifyFileSignatureWorkflow( context_r ) );
-      return context_r.fileAccepted();
-    }
-    bool _verifyFileSignatureWorkflow( keyring::VerifyFileContext & context_r );
-
-    bool verifyFileSignature( const Pathname & file, const Pathname & signature )
-    { return verifyFile( file, signature, generalKeyRing() ); }
-    bool verifyFileTrustedSignature( const Pathname & file, const Pathname & signature )
-    { return verifyFile( file, signature, trustedKeyRing() ); }
-
-    PublicKeyData publicKeyExists( const std::string & id )
-    { return publicKeyExists(id, generalKeyRing());}
-    PublicKeyData trustedPublicKeyExists( const std::string & id )
-    { return publicKeyExists(id, trustedKeyRing());}
-
-    bool provideAndImportKeyFromRepositoryWorkflow (const std::string &id_r , const RepoInfo &info_r );
-
-    void allowPreload( bool yesno_r )
-    { _allowPreload = yesno_r; }
-
-  private:
-    /** Impl helper providing on demand a KeyManagerCtx to manip a cached keyring. */
-    CachedPublicKeyData::Manip keyRingManip( const Pathname & keyring )
-    { return cachedPublicKeyData.manip( keyring ); }
-
-    bool verifyFile( const Pathname & file, const Pathname & signature, const Pathname & keyring );
-    void importKey( const Pathname & keyfile, const Pathname & keyring );
-
-    PublicKey exportKey( const std::string & id, const Pathname & keyring );
-    PublicKey exportKey( const PublicKeyData & keyData, const Pathname & keyring );
-    PublicKey exportKey( const PublicKey & key, const Pathname & keyring )
-    { return exportKey( key.keyData(), keyring ); }
-
-    void dumpPublicKey( const std::string & id, const Pathname & keyring, std::ostream & stream );
-    filesystem::TmpFile dumpPublicKeyToTmp( const std::string & id, const Pathname & keyring );
-
-    void deleteKey( const std::string & id, const Pathname & keyring );
-
-    std::list<PublicKey> publicKeys( const Pathname & keyring);
-    const std::list<PublicKeyData> & publicKeyData( const Pathname & keyring )
-    { return cachedPublicKeyData( keyring ); }
-
-    /** Get \ref PublicKeyData for ID (\c false if ID is not found). */
-    PublicKeyData publicKeyExists( const std::string & id, const Pathname & keyring );
-    /** Load key files cached on the system into the generalKeyRing. */
-    void preloadCachedKeys();
-
-    const Pathname generalKeyRing() const
-    { return _general_tmp_dir.path(); }
-    const Pathname trustedKeyRing() const
-    { return _trusted_tmp_dir.path(); }
-
-    // Used for trusted and untrusted keyrings
-    filesystem::TmpDir _trusted_tmp_dir;
-    filesystem::TmpDir _general_tmp_dir;
-    Pathname _base_dir;
-    bool _allowPreload = false;	//< General keyring may be preloaded with keys cached on the system.
-
-  private:
-    /** Functor returning the keyrings data (cached).
-     * \code
-     *  const std::list<PublicKeyData> & cachedPublicKeyData( const Pathname & keyring );
-     * \endcode
-     */
-    CachedPublicKeyData cachedPublicKeyData;
-  };
   ///////////////////////////////////////////////////////////////////
 
   namespace
@@ -362,6 +207,13 @@ namespace zypp
     };
   } // namespace
 
+  KeyRing::Impl::Impl(const filesystem::Pathname &baseTmpDir)
+    : _trusted_tmp_dir( baseTmpDir, "zypp-trusted-kr" )
+    , _general_tmp_dir( baseTmpDir, "zypp-general-kr" )
+    , _base_dir( baseTmpDir )
+  {
+    MIL << "Current KeyRing::DefaultAccept: " << _keyRingDefaultAccept << std::endl;
+  }
 
   void KeyRing::Impl::importKey( const PublicKey & key, bool trusted )
   {
